requirements.txt
numpy==2.3.0
pandas==2.3.0
fastapi==0.115.12
uvicorn==0.34.3
google-cloud-firestore==2.21.0